Re-arrange hours functions to hours.go
